+++ src/main/java/com/kortekslab/location/service/ICampusService.java
@@ -8,5 +8,5 @@ public interface ICampusService {
 	public Campus addCampus(Campus campus);
 	public Campus updateCampus(Campus campus);
 	public List<Campus> findAll();
-	public Campus findFirstById(int id);
+	public Campus findFirstById(long id);
 }
